Failed attempt at Day 19 part 2
Tries to reduce the target molecule back down to "e" by randomly selecting and applying one of the replacement rules that is applicable to the current state of the molecule; and starting over from the beginning if it hits a dead end.

This ran slowly and didn't produce a solution for me.